Rename 폴더 변경, 미사용 import 삭제
