+++ src/firebase.js
@@ -1,8 +1,6 @@
 // firebase.js
-
 // Import the functions you need from the SDKs you need
 import { initializeApp } from 'firebase/app';
-
 import {
   createUserWithEmailAndPassword,
   getAuth,
